refactor: make PopUp component handle both booka and genre deletion

+++ client/src/components/PopUp.jsx
@@ -1,4 +1,12 @@
-const PopUp = ({ isOpen, onClose, password, onPasswordChange, deleteBook }) => {
+const PopUp = ({
+  isOpen,
+  onClose,
+  password,
+  onPasswordChange,
+  deleteBook,
+  deleteGenre,
+  itemToDelete,
+}) => {
   if (!isOpen) {
     return null;
   }
@@ -39,9 +47,13 @@ const PopUp = ({ isOpen, onClose, password, onPasswordChange, deleteBook }) => {
           <button
             className="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700"
             onClick={() => {
-              deleteBook();
+              if (itemToDelete.type === "genre") {
+                deleteGenre();
+              }
+              if (itemToDelete.type === "book") {
+                deleteBook();
+              }
               onClose();
-              () => console.log("confirm button clicked");
             }}
           >
             Confirm
